requirements.txt
discord.py[speed]==2.3.2
humanize>=4.8.0,<5
jishaku>=2.5.2,<3
Pillow>=10.0.0,<11
redis[hiredis]>=5.0.1,<6
rich==13.6.0,<14
uvloop==0.19.0,<0.20